refactor: [포트폴리오] main.js 파일 모듈화
- 기존 main.js 파일 내의 중복된 스크립트 제거
- 한 파일 내의 길었던 파일 기능에 따라 모듈화
