+++ main.html
@@ -8,7 +8,7 @@
   <link rel="stylesheet" href="./src/style.css">
   <link rel="stylesheet" href="./src/main.css">
   <script src="https://cdnjs.cloudflare.com/ajax/libs/html2canvas/1.4.1/html2canvas.min.js"></script>
-  <script src="./src/main.js" defer></script>
+  <script src="./src/main.js" type="module"></script>
 </head>
 
 <body>
